refactor(*): Refactor components
